+++ cpp/open3d/Open3D.h.in
@@ -66,9 +66,6 @@
 #include "open3d/utility/FileSystem.h"
 #include "open3d/utility/Helper.h"
 #include "open3d/utility/Timer.h"
-#include "open3d/visualization/Utility/DrawGeometry.h"
-#include "open3d/visualization/Utility/SelectionPolygon.h"
-#include "open3d/visualization/Utility/SelectionPolygonVolume.h"
 #include "open3d/visualization/gui/Application.h"
 #include "open3d/visualization/gui/Button.h"
 #include "open3d/visualization/gui/Checkbox.h"
@@ -87,6 +84,9 @@
 #include "open3d/visualization/gui/TextEdit.h"
 #include "open3d/visualization/gui/Theme.h"
 #include "open3d/visualization/gui/Window.h"
+#include "open3d/visualization/utility/DrawGeometry.h"
+#include "open3d/visualization/utility/SelectionPolygon.h"
+#include "open3d/visualization/utility/SelectionPolygonVolume.h"
 #include "open3d/visualization/visualizer/ViewControl.h"
 #include "open3d/visualization/visualizer/ViewControlWithCustomAnimation.h"
 #include "open3d/visualization/visualizer/ViewControlWithEditing.h"
