+++ cpp/open3d/io/sensor/azure_kinect/AzureKinectRecorder.cpp
@@ -42,7 +42,7 @@
 #include "open3d/geometry/RGBDImage.h"
 #include "open3d/io/sensor/azure_kinect/K4aPlugin.h"
 #include "open3d/io/sensor/azure_kinect/MKVReader.h"
-#include "open3d/visualization/Utility/ColorMap.h"
+#include "open3d/visualization/utility/ColorMap.h"
 #include "open3d/visualization/visualizer/VisualizerWithKeyCallback.h"
 
 namespace open3d {
