+++ cpp/open3d/visualization/Shader/ImageShader.cpp
@@ -30,7 +30,7 @@
 
 #include "open3d/geometry/Image.h"
 #include "open3d/visualization/Shader/Shader.h"
-#include "open3d/visualization/Utility/ColorMap.h"
+#include "open3d/visualization/utility/ColorMap.h"
 
 namespace open3d {
 namespace visualization {
