+++ cpp/open3d/visualization/utility/PointCloudPicker.cpp
@@ -24,7 +24,7 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/visualization/Utility/PointCloudPicker.h"
+#include "open3d/visualization/utility/PointCloudPicker.h"
 
 #include "open3d/geometry/BoundingVolume.h"
 #include "open3d/geometry/PointCloud.h"
