+++ cpp/open3d/visualization/visualizer/VisualizerWithVertexSelection.cpp
@@ -38,10 +38,10 @@
 #include "open3d/io/PointCloudIO.h"
 #include "open3d/io/TriangleMeshIO.h"
 #include "open3d/utility/FileSystem.h"
-#include "open3d/visualization/Utility/GLHelper.h"
-#include "open3d/visualization/Utility/PointCloudPicker.h"
-#include "open3d/visualization/Utility/SelectionPolygon.h"
-#include "open3d/visualization/Utility/SelectionPolygonVolume.h"
+#include "open3d/visualization/utility/GLHelper.h"
+#include "open3d/visualization/utility/PointCloudPicker.h"
+#include "open3d/visualization/utility/SelectionPolygon.h"
+#include "open3d/visualization/utility/SelectionPolygonVolume.h"
 #include "open3d/visualization/visualizer/RenderOptionWithEditing.h"
 #include "open3d/visualization/visualizer/ViewControlWithEditing.h"
