+++ cpp/tests/geometry/VoxelGrid.cpp
@@ -27,7 +27,7 @@
 #include "open3d/geometry/VoxelGrid.h"
 #include "open3d/geometry/LineSet.h"
 #include "open3d/geometry/TriangleMesh.h"
-#include "open3d/visualization/Utility/DrawGeometry.h"
+#include "open3d/visualization/utility/DrawGeometry.h"
 #include "tests/UnitTest.h"
 
 namespace open3d {
