+++ cpp/tests/io/ClassIO/VoxelGridIO.cpp
@@ -26,7 +26,7 @@
 
 #include "open3d/io/VoxelGridIO.h"
 #include "open3d/geometry/VoxelGrid.h"
-#include "open3d/visualization/Utility/DrawGeometry.h"
+#include "open3d/visualization/utility/DrawGeometry.h"
 #include "tests/UnitTest.h"
 
 namespace open3d {
